remove sample data and unused components
