+++ app/src/pages/ComponentDemos.tsx
@@ -1,0 +1,110 @@
+import { Button } from "@/components/ui/button";
+import { Calendar } from "@/components/ui/calendar/calendar";
+import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import React from "react";
+import { DatePickerDemo } from "./DatePickerDemo";
+import { SidebarTrigger } from "@/components/ui/sidebar";
+import { DataTable } from "@/components/table/Table";
+import type { ColumnDef } from "@tanstack/react-table";
+
+function Demo() {
+  const [date, setDate] = React.useState<Date | undefined>(new Date())
+
+  //Dev note: these types we will need to define on the components where we want to render the table, probably in its own file is good
+
+  type Animal = {
+    id: string;
+    animalName: string;
+    age: number;
+    gender: "Male" | "Female";
+    status: "Adopted" | "Available";
+  };
+
+  //dev note these need to match the type
+
+  const columns: ColumnDef<Animal>[] = [
+    { accessorKey: "id", header: "Animal ID" },
+    {
+      accessorKey: "animalName",
+      header: "Animal Name",
+    },
+    {
+      accessorKey: "age",
+      header: "Age",
+    },
+    {
+      accessorKey: "gender",
+      header: "Gender",
+    },
+    {
+      accessorKey: "status",
+      header: "Status",
+    },
+  ];
+
+  const data: Animal[] = [
+    { id: "1", animalName: "Lucky", age: 2, gender: "Male", status: "Adopted" },
+    {
+      id: "2",
+      animalName: "Spot",
+      age: 2,
+      gender: "Female",
+      status: "Adopted",
+    },
+    {
+      id: "3",
+      animalName: "Rover",
+      age: 4,
+      gender: "Male",
+      status: "Available",
+    },
+    { id: "4", animalName: "Rex", age: 3, gender: "Male", status: "Adopted" },
+    {
+      id: "5",
+      animalName: "Chuck",
+      age: 7,
+      gender: "Male",
+      status: "Available",
+    },
+  ];
+
+  return (
+    <>
+        <div className="flex flex-1 flex-col gap-4 p-4 items-center">
+
+          <Button>default</Button>
+          <Button variant="outline">outline</Button>
+          <Button variant="secondary">secondary</Button>
+          <br />
+          <Select>
+            <SelectTrigger className="w-[180px]">
+              <SelectValue placeholder="Select a role" />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectGroup>
+                <SelectItem value="fosterparent">Foster Parent</SelectItem>
+                <SelectItem value="volunteer">Volunteer</SelectItem>
+                <SelectItem value="employee">Employee</SelectItem>
+              </SelectGroup>
+            </SelectContent>
+          </Select>
+          <br />
+
+          <Calendar
+            mode="single"
+            selected={date}
+            onSelect={setDate}
+            className="rounded-lg border"
+          />
+
+          <DatePickerDemo />
+
+          <div className="container mx-auto py-10">
+            <DataTable columns={columns} data={data} />
+          </div>
+        </div>
+      </>
+  );
+}
+
+export default Demo;
