--- app/src/pages/Dashboard.tsx
@@ -1,9 +1,0 @@
-
-
-function Dashboard() {
-return (
-<></>
-  )
-}
-
-export default Dashboard;
